requirements.txt
# NYXMed VFinal - Production Pipeline
# =====================================

# Core
requests>=2.28.0
numpy>=1.21.0
pandas>=1.4.0

# RAG Components
faiss-cpu>=1.7.3
rank-bm25>=0.2.2
sentence-transformers>=2.2.0

# AWS S3 (for index building from cloud data)
boto3>=1.26.0

# Optional: OpenAI fallback
openai>=1.0.0

# Optional: GPU support
# faiss-gpu>=1.7.3

# Development
python-dotenv>=1.0.0
